Fix data types and original epoch time

+++ core/src/main/java/io/github/cactacea/core/domain/enums/AccountStatusType.java
@@ -3,17 +3,17 @@ package io.github.cactacea.core.domain.enums;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public enum AccountStatusType {
-    @JsonProperty("0") singedUp(0),
-    @JsonProperty("1") deleted(1),
-    @JsonProperty("2") terminated(2);
+    @JsonProperty("0") singedUp((byte)0),
+    @JsonProperty("1") deleted((byte)1),
+    @JsonProperty("2") terminated((byte)2);
 
-    private long value;
+    private byte value;
 
-    private AccountStatusType(long value) {
+    private AccountStatusType(byte value) {
         this.value = value;
     }
 
-    static public AccountStatusType forName(long value) {
+    static public AccountStatusType forName(byte value) {
         for (AccountStatusType e : values()) {
             if (e.value == value) {
                 return e;
@@ -22,7 +22,7 @@ public enum AccountStatusType {
         throw new IllegalArgumentException();
     }
 
-    public long toValue() {
+    public byte toValue() {
         return value;
     }
 }
